implement dynamic code and display the boards in the front
