+++ .idea/.gitignore
@@ -7,3 +7,4 @@
 /dataSources/
 /dataSources.local.xml
  node_moudules
+ view engine
