+++ assests/css/Style.css
@@ -11,18 +11,18 @@ body {
     background: #f5f5f5;
     font-family: 'Varela Round', sans-serif;
     font-size: 14px;
-
+    
 }
 /*******navbar*/
 .menu
 {
-    background-color:  #435d7d;
-    border-bottom: 4px solid #79b8d9;
-    width:100%;
-    height: auto;
-    padding: 0 10px;
+background-color:  #435d7d;
+border-bottom: 4px solid #79b8d9;
+width:100%;
+height: auto;
+padding: 0 10px;
 
-    margin: 0px;
+margin: 0px;
 
 
 }
@@ -34,38 +34,38 @@ body {
 
 .menu  .navbar-nav > .active > a,h2
 {
-    /*background-color : #04A3ED;*/
-    margin-top: 0px;
-    margin-bottom: 0px;
-    color: white;
-    font-weight: bold;
+/*background-color : #04A3ED;*/
+margin-top: 0px;
+margin-bottom: 0px;
+color: white;
+font-weight: bold;
 }
 
 .menu  .navbar-nav >  li >  a
 {
-    font-size: 14px;
-    color: white;
-    padding: 10px 35px;
+font-size: 14px;
+color: white;
+padding: 10px 35px;
 
 }
 .menu  .navbar-nav >  li >  a:hover
 {
-    background-color: #04A3ED;
+background-color: #04A3ED;
 }
 
 .navbar-header > a
 {
-    font-family: 'Ubuntu Condensed', sans-serif;
-    padding: 0px;
-    margin: 0px;
-    text-decoration: none;
-    color: white;
-    font-size: 25px;
-    padding: 5px 30px;
+font-family: 'Ubuntu Condensed', sans-serif;
+padding: 0px;
+margin: 0px;
+text-decoration: none;
+color: white;
+font-size: 25px;
+padding: 5px 30px;
 }
 .navbar-header > a:hover
 {
-    text-decoration: none;
+text-decoration: none;
 
 
 }
@@ -75,13 +75,13 @@ body {
     margin: 50px 0;
 }
 .table-wrapper {
-    width: auto;
+   width: auto;
     background: #fff;
     padding: 20px 25px;
     border-radius: 3px;
     box-shadow: 0 1px 1px rgba(50, 41, 41, 0.05);
 }
-.table-title {
+.table-title {        
     padding-bottom: 15px;
     background: #435d7d;
     color: #fff;
@@ -114,7 +114,7 @@ body {
 }
 .clearfix{
     margin-left: 50px;
-
+    
 }
 .clearfix a {
     display: block;
@@ -144,8 +144,8 @@ table.table tr th, table.table tr td {
     padding: 12px 15px;
     vertical-align: middle;
     text-align: center;
-
-
+    
+   
 }
 
 #td-id-3{
@@ -169,7 +169,7 @@ table.table th i {
     font-size: 13px;
     margin: 0 5px;
     cursor: pointer;
-}
+}	
 table.table td:last-child i {
     opacity: 0.9;
     font-size: 22px;
@@ -183,13 +183,13 @@ table.table td a {
     outline: none !important;
 }
 /*
-tr:hover {
-    background: red;
+tr:hover { 
+    background: red; 
  }
- td a {
-    display: block;
+ td a { 
+    display: block; 
     border: 1px solid black;
-    padding: 16px;
+    padding: 16px; 
  }
  */
 table.table td span .text-success{
@@ -251,11 +251,11 @@ table.table .avatar {
 }
 .pagination li a:hover {
     color: #666;
-}
+}	
 .pagination li.active a, .pagination li.active a.page-link {
     background: #03A9F4;
 }
-.pagination li.active a:hover {
+.pagination li.active a:hover {        
     background: #0397d6;
 }
 .pagination li.disabled i {
@@ -269,12 +269,12 @@ table.table .avatar {
     float: left;
     margin-top: 10px;
     font-size: 13px;
-}
+}    
 /* Custom checkbox */
 .custom-checkbox {
     position: relative;
 }
-.custom-checkbox input[type="checkbox"] {
+.custom-checkbox input[type="checkbox"] {    
     opacity: 0;
     position: absolute;
     margin: 5px 0 0 3px;
@@ -349,7 +349,7 @@ table.table .avatar {
 .modal .btn {
     border-radius: 2px;
     min-width: 100px;
-}
+}	
 .modal form label {
     font-weight: normal;
 }
