+++ package.json
@@ -5,7 +5,8 @@
   "main": "BoardStatisticObj.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "dev": "nodemon server.js"
   },
   "repository": {
     "type": "git",
@@ -18,7 +19,11 @@
   },
   "homepage": "https://github.com/GavrieliYovel/Core-Team#readme",
   "dependencies": {
+    "axios": "^1.2.0",
+    "ejs": "^3.1.8",
     "express": "^4.18.2",
-    "moment": "^2.29.4"
+    "json-2-csv": "^3.18.0",
+    "moment": "^2.29.4",
+    "nodemon": "^2.0.20"
   }
 }
