+++ views/boards.ejs
@@ -1,0 +1,226 @@
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	
+<meta charset="utf-8">
+<meta http-equiv="X-UA-Compatible" content="IE=edge">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<title>Task Manger</title>
+<!-- Boostrap Links-->
+<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto|Varela+Round">
+<link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
+<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
+
+
+<!------ Include the above in your HEAD tag ---------->
+
+ <!-- Style css -->
+<link rel="stylesheet" href="http://localhost:8080/assests/css/Style.css" />
+<script>
+	$(document).ready(function(){
+		// Activate tooltip
+		$('[data-toggle="tooltip"]').tooltip();
+
+		// Select/Deselect checkboxes
+		var checkbox = $('table tbody input[type="checkbox"]');
+		$("#selectAll").click(function(){
+			if(this.checked){
+				checkbox.each(function(){
+					this.checked = true;
+				});
+			} else{
+				checkbox.each(function(){
+					this.checked = false;
+				});
+			}
+		});
+		checkbox.click(function(){
+			if(!this.checked){
+				$("#selectAll").prop("checked", false);
+			}
+		});
+		$(function()
+				{
+					$('#table').bootstrapTable()
+				}
+		)
+
+		//Moment.JS Return Date Ranges
+		function getDates(startDate, stopDate) {
+			var dateArray = [];
+			var currentDate = moment(startDate);
+			var stopDate = moment(stopDate);
+			while (currentDate <= stopDate) {
+				dateArray.push( moment(currentDate).format('YYYY-MM-DD'))
+				currentDate = moment(currentDate).add(1, 'days');
+			}
+			return dateArray;
+		}
+
+
+		$('#ok').click( function()
+		{
+
+			var $table = $('#table')
+			var from=$("input[type=date][name=date1]" ).val();
+			var to=$("input[type=date][name=date2]" ).val();
+			$table.bootstrapTable('filterBy',{ ETA: getDates(from,to)})
+
+
+
+		})
+	});
+$(function(){
+     $(".table").on("click", "td[role=\"button\"]", function (e) {
+          window.location = $(this).data("href");
+     });
+});
+</script>
+</head>
+<body>
+	<div class="menu">
+		<div class="container-fluid">
+			<div class="navbar-header">
+				<h2>Task<b>Manger</b></h2>
+			</div>
+			<div>
+				<ul class="nav navbar-nav navbar-right">
+
+					<li><a href="#"><span class="glyphicon glyphicon-log-in"></span> Login</a></li>
+				</ul>
+			</div>
+		</div>
+	</div>
+    <div class="container">
+		<div class="table-responsive">
+			<div class="table-wrapper">
+				<div class="table-title">
+					<div class="row">
+						<div class="col-xs-6">
+							<h2>Boards</h2>
+						</div>
+						<div class="col-xs-6">
+							<a href="#addBoardModal" class="btn btn-success" data-toggle="modal"><i class="material-icons">&#xE147;</i> <span>Add New Board</span></a>
+							<a href="#" class="btn btn-primary"><i class="material-icons">&#xE24D;</i> <span>Export to CSV</span></a>
+							<a href="#deleteBoardModal" class="btn btn-danger" data-toggle="modal"><i class="material-icons">&#xE15C;</i> <span>Delete</span></a>
+						</div>
+					</div>
+				</div>
+				<table class="table table-striped table-hover">
+					<thead>
+						<tr>
+<!--							<th>-->
+<!--								<span class="custom-checkbox">-->
+<!--									<input type="checkbox" id="selectAll">-->
+<!--									<label for="selectAll"></label>-->
+<!--								</span>-->
+<!--							</th>-->
+
+							<th>BoardsID</th>
+							<th>BoardsName</th>
+						</tr>
+					</thead>
+					<tbody id="tbody" class="tbody">
+
+					<%- include('include/_show') %>
+
+					</tbody>
+				</table>
+
+			</div>
+		</div>
+    </div>
+	<!-- Edit Modal HTML -->
+	<div id="addBoardModal" class="modal fade">
+		<div class="modal-dialog">
+			<div class="modal-content">
+				<form action="/api/boards" method="POST" id="boards">
+					<div class="modal-header">
+						<h4 class="modal-title">Add Boards</h4>
+						<button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+					</div>
+					<div class="modal-body">
+						<div class="form-group">
+							<label>BoardsID</label>
+							<input type="text" class="form-control" required>
+						</div>
+						<div class="form-group">
+							<label>BoardsName</label>
+							<input type="text" class="form-control" required>
+						</div>
+					</div>
+					<div class="modal-footer">
+						<input type="button" class="btn btn-default" data-dismiss="modal" value="Cancel">
+						<input type="submit" class="btn btn-success" value="Add">
+					</div>
+				</form>
+			</div>
+		</div>
+	</div>
+	<!-- Edit Modal HTML -->
+	<div id="editEmployeeModal" class="modal fade">
+		<div class="modal-dialog">
+			<div class="modal-content">
+				<form>
+					<div class="modal-header">
+						<h4 class="modal-title">Edit Boards</h4>
+						<button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+					</div>
+					<div class="modal-body">
+						<div class="form-group">
+							<label>IDBoards</label>
+							<input type="text" class="form-control" required>
+						</div>
+						<div class="form-group">
+							<label>BoardsName</label>
+							<input type="text" class="form-control" required>
+						</div>
+					</div>
+					<div class="modal-footer">
+						<input type="button" class="btn btn-default" data-dismiss="modal" value="Cancel">
+						<input type="submit" class="btn btn-info" value="Save">
+					</div>
+				</form>
+			</div>
+		</div>
+	</div>
+	<!-- Delete Modal HTML -->
+	<div id="deleteBoardModal" class="modal fade">
+		<div class="modal-dialog">
+			<div class="modal-content">
+				<form>
+					<div class="modal-header">
+						<h4 class="modal-title">Delete Boards</h4>
+						<button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+					</div>
+					<div class="modal-body">
+						<p>Are you sure you want to delete these Records?</p>
+						<p class="text-warning"><small>This action cannot be undone.</small></p>
+					</div>
+					<div class="modal-footer">
+						<input type="button" class="btn btn-default" data-dismiss="modal" value="Cancel">
+						<input type="submit" class="btn btn-danger" value="Delete">
+					</div>
+				</form>
+			</div>
+		</div>
+	</div>
+<!--	<script>-->
+<!--		var sendRequest = function () {-->
+
+<!--		fetch("http://localhost:8080/api/boards")-->
+<!--			.then((data)=>{-->
+<!--			console.log(data);-->
+<!--			return data.json();-->
+<!--			}).then((objectData)=> {-->
+<!--			console.log(JSON.stringify(parse(objectData)));-->
+<!--			}).catch(error => console.log('error', error));-->
+<!--			}-->
+<!--	</script>-->
+<!--	<script> window.onload=sendRequest()</script>-->
+</body>
+</html>
