+++ views/include/_cerateBoard.ejs
@@ -1,0 +1,26 @@
+<div id="addBoardModal" class="modal fade">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <form action="/api/boards" method="POST" id="boards">
+                <div class="modal-header">
+                    <h4 class="modal-title">Add Boards</h4>
+                    <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
+                </div>
+                <div class="modal-body">
+                    <div class="form-group">
+                        <label>BoardsID</label>
+                        <input type="text" class="form-control" required>
+                    </div>
+                    <div class="form-group">
+                        <label>BoardsName</label>
+                        <input type="text" class="form-control" required>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <input type="button" class="btn btn-default" data-dismiss="modal" value="Cancel">
+                    <input type="submit" class="btn btn-success" value="Add">
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
