+++ views/tasks.ejs
@@ -18,7 +18,8 @@
 	<!------ Include the above in your HEAD tag ---------->
 
 	<!-- Style css -->
-	<link rel="stylesheet" href="../assests/css/Style.css" />
+	<link rel="stylesheet" href="http://localhost:8080/assests/css/Style.css" />
+
 	<script>
 		$(document).ready(function(){
 			// Activate tooltip
@@ -427,7 +428,7 @@
 	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
 	  integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
 	  crossorigin="anonymous"></script>
-	  <script src="../assests/js/tasks.js"></script>
+	  <script src="./tasks.js"></script>
 
 </body>
 </html>
